Update setup.py
Change setup.py to use the correct method of specifying a required
python version (requirements here are: greater than 2.6 & not 3.2)

Previously specified the python version as one element in the
`requires` argument, I'm not sure if this method was ever official
supported, and if it was, it is deprecated now, and causes problems
with other tools, e.g. pex.

(Please see: pantsbuild/pex#1201 for an
example of the kind of error this throws)

this commit updates the arguments to setup.py to use the now standard
python_requires keyword argument to specify the required python version